requirements.txt
# Core PyTorch and GPU Libraries (Essential for GPU Training)
# Note: Usually, installing torch with the correct CUDA version pulls these in.
torch==2.2.0
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvcc-cu12==12.6.85 # Needed if compiling CUDA code (e.g., some optim kernels)
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.6.20
# nvidia-nvtx-cu12==12.1.105

# PyTorch Lightning (High-level training wrapper) & Utilities
pytorch-lightning==2.4.0
lightning-utilities==0.11.6
torchmetrics==1.4.1         # Common metrics library used with PyTorch/Lightning

# Transformer / Model Building Blocks
einops==0.8.0               # Often used for tensor manipulation in transformers
# flash-attn==2.7.2.post1     # Efficient attention implementation (GPU)
ninja==1.11.1.3             # Often needed for building custom CUDA ops (like flash-attn)

# Weights & Biases (Experiment Tracking)
wandb==0.17.7
# Core wandb dependencies (often pulled automatically, but good to list):
PyYAML==6.0.2
requests==2.31.0
psutil==6.0.0              
protobuf==5.27.3
setproctitle==1.3.3
GitPython==3.1.43          # Used by wandb to track code state
docker-pycreds==0.4.0      # Used by wandb for system metrics/docker integration

# Core Data Handling & Numerics
numpy==1.26.4
pandas==2.2.2            

# Common Utilities & Dependencies (often required by the above)
packaging==23.2
tqdm==4.65.0                # Progress bars
PyYAML==6.0.2               # Already listed (wandb), also often used for configs
typing_extensions==4.12.2   # Required by many modern libraries (like pydantic, used by wandb/PL)
pydantic==2.8.2             # Data validation library, dependency of wandb/PL
pydantic_core==2.20.1       # Core part of pydantic
filelock==3.15.4            # Used by torch, huggingface libs
sympy==1.13.2               # Used by torch
networkx==3.3               # Used by torch
jinja2==3.1.4               # Used by torch, PL
MarkupSafe==3.0.2           # Dependency of Jinja2
fsspec==2024.6.1            # Filesystem interactions, used by PL/wandb

# Optional: TensorBoard Logging (Often used alongside or default in Pytorch Lightning)
tensorboard==2.18.0
tensorboard-data-server==0.7.2
grpcio==1.67.1              # Dependency for TensorBoard
